Separate git and git-prep-directory
The idea is to more clearly separate what are just git wrapper functions
and what is introduced due to git-prep-directory.
