+++ git.go
@@ -3,10 +3,8 @@ package git
 import (
 	"fmt"
 	"io"
-	"log"
 	"os"
 	"os/exec"
-	"path"
 	"path/filepath"
 	"regexp"
 	"strings"
@@ -81,6 +79,30 @@ func ContextRun(ctx context.Context, cmd *exec.Cmd) error {
 	case err := <-errc:
 		return err // err may be nil
 	}
+}
+
+// Checkout switches branches or restores working tree files.
+func Checkout(gitDir, checkoutDir, ref string) error {
+	err := os.MkdirAll(checkoutDir, 0777)
+	if err != nil {
+		return err
+	}
+
+	args := []string{"--work-tree", checkoutDir, "checkout", ref, "--", "."}
+	err = Command(gitDir, "git", args...).Run()
+	if err != nil {
+		return err
+	}
+
+	// Set mtimes to time file is most recently affected by a commit.
+	// This is annoying but unfortunately git sets the timestamps to now,
+	// and docker depends on the mtime for cache invalidation.
+	err = SetMTimes(gitDir, checkoutDir, ref)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
 
 // Fetch fetches all branches from a given remote.
@@ -166,89 +188,6 @@ func Describe(gitDir, ref string) (desc string, err error) {
 	return desc, nil
 }
 
-// Checkout switches branches or restores working tree files.
-func Checkout(gitDir, checkoutDir, ref string) error {
-	err := os.MkdirAll(checkoutDir, 0777)
-	if err != nil {
-		return err
-	}
-
-	args := []string{"--work-tree", checkoutDir, "checkout", ref, "--", "."}
-	err = Command(gitDir, "git", args...).Run()
-	if err != nil {
-		return err
-	}
-
-	// Set mtimes to time file is most recently affected by a commit.
-	// This is annoying but unfortunately git sets the timestamps to now,
-	// and docker depends on the mtime for cache invalidation.
-	err = SetMTimes(gitDir, checkoutDir, ref)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// BuildDirectory holds the git rev and path to a cloned git repository. It also
-// holds a cleanup function to safely remove this directory.
-type BuildDirectory struct {
-	Name    string
-	Dir     string
-	Cleanup func()
-}
-
-// PrepBuildDirectory clones a given repository and checks out the given
-// revision, setting the timestamp of all files to their commit time and putting
-// all submodules into a submodule cache.
-func PrepBuildDirectory(gitDir, remote, ref string) (*BuildDirectory, error) {
-	start := time.Now()
-	defer func() {
-		log.Printf("Took %v to prep %v", time.Since(start), remote)
-	}()
-
-	if strings.HasPrefix(remote, "github.com/") {
-		remote = "https://" + remote
-	}
-
-	gitDir, err := filepath.Abs(gitDir)
-	if err != nil {
-		return nil, fmt.Errorf("unable to determine abspath: %v", err)
-	}
-
-	err = LocalMirror(remote, gitDir, ref, os.Stderr)
-	if err != nil {
-		return nil, fmt.Errorf("unable to LocalMirror: %v", err)
-	}
-
-	rev, err := RevParse(gitDir, ref)
-	if err != nil {
-		return nil, fmt.Errorf("unable to parse rev: %v", err)
-	}
-
-	tagName, err := Describe(gitDir, rev)
-	if err != nil {
-		return nil, fmt.Errorf("unable to describe %v: %v", rev, err)
-	}
-
-	shortRev := rev[:10]
-	checkoutPath := path.Join(gitDir, filepath.Join("c/", shortRev))
-
-	err = RecursiveCheckout(gitDir, checkoutPath, rev)
-	if err != nil {
-		return nil, err
-	}
-
-	cleanup := func() {
-		err := SafeCleanup(checkoutPath)
-		if err != nil {
-			log.Println("Error cleaning up path:", checkoutPath)
-		}
-	}
-
-	return &BuildDirectory{tagName, checkoutPath, cleanup}, nil
-}
-
 // RecursiveCheckout recursively checks out repositories; similar to "git clone
 // --recursive".
 func RecursiveCheckout(gitDir, checkoutPath, rev string) error {
@@ -262,13 +201,4 @@ func RecursiveCheckout(gitDir, checkoutPath, rev string) error {
 		return fmt.Errorf("failed to prep submodules: %v", err)
 	}
 	return nil
-}
-
-// SafeCleanup recursively removes all files from a given path, which has to be
-// a subfolder of the current working directory.
-func SafeCleanup(path string) error {
-	if path == "/" || path == "" || path == "." || strings.Contains(path, "..") {
-		return fmt.Errorf("invalid path specified for deletion %q", path)
-	}
-	return os.RemoveAll(path)
 }
